backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pre-commit==3.8.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
PyJWT==2.8.0
email_validator==2.2.0
passlib[bcrypt]<2.0.0,>=1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
python-json-logger==2.0.7
asyncpg==0.30.0
structlog==24.4.0
stripe==11.4.1
gunicorn==23.0.0
alembic-postgresql-enum==1.7.0
jinja2==3.1.3
apscheduler==3.11.0
aioboto3==13.3.0
httpx